Updated the RTS sample to show proper UI (#1700)
* Updated the RTS sample to show proper UI

The RTS sample UI is now fully done with RML UI.
A default RML style has been added, to make it easier to get started with UI in other projects.
The sample now also shows a main menu (press ESC to show).
The main menu has a few real options (such as UI scaling) and a number
of example UI elements to show how to use them.
The sample window is now also resizable to demonstrate the UI scaling.

* materials