Mesh particles now use sub-meshes for variation (#1726)
If a mesh particle renderer references a mesh that has multiple sub-meshes, the sub-meshes are randomly selected for the particle. This way you can put multiple meshes for, e.g. broken pieces, into a mesh to have more variety in the effect.

Also added a scale parameter to the mesh renderer, to adjust the size of the mesh independently of the particle size.